+++ antlr/SysYListener.cpp
@@ -1,7 +1,0 @@
-
-// Generated from D:/SELF_YANG/appdata/IDEA/repos/test/SysY.g4 by ANTLR 4.13.2
-
-
-#include "SysYListener.h"
-
-